Moved sign up confirmation into sign up page #1866
