+++ src/Tooling/Bit.Tooling.Templates/TodoTemplate/Web/Pages/SignUp.razor.cs
@@ -7,6 +7,7 @@ public partial class SignUp
 {
     public SignUpModel SignUpModel { get; set; } = new();
 
+    public bool IsSignedUp { get; set; }
     public bool IsLoading { get; set; }
 
     public BitMessageBarType SignUpMessageType { get; set; }
@@ -38,7 +39,10 @@ public partial class SignUp
                 Password = SignUpModel.Password
             }, TodoTemplateJsonContext.Default.SignUpRequestDto);
 
-            NavigationManager.NavigateTo($"/sign-up-confirmation?email={SignUpModel.Email}");
+            IsSignedUp = true;
+
+            SignUpMessageType = BitMessageBarType.Success;
+            SignUpMessage = "A confirmation link has been sent to your email.";
         }
         catch (ResourceValidationException e)
         {
@@ -59,6 +63,39 @@ public partial class SignUp
         }
     }
 
+    private async Task ResendLink()
+    {
+        if (IsLoading)
+        {
+            return;
+        }
+
+        IsLoading = true;
+        SignUpMessage = null;
+
+        try
+        {
+            await HttpClient.PostAsJsonAsync("Auth/SendConfirmationEmail", new()
+            {
+                Email = SignUpModel.Email
+            }, TodoTemplateJsonContext.Default.SendConfirmationEmailRequestDto);
+
+            SignUpMessageType = BitMessageBarType.Success;
+
+            SignUpMessage = "The confirmation link has been re-sent.";
+        }
+        catch (KnownException e)
+        {
+            SignUpMessageType = BitMessageBarType.Error;
+
+            SignUpMessage = ErrorStrings.ResourceManager.Translate(e.Message, SignUpModel.Email!);
+        }
+        finally
+        {
+            IsLoading = false;
+        }
+    }
+
     protected override async Task OnAfterRenderAsync(bool firstRender)
     {
         await base.OnAfterRenderAsync(firstRender);
@@ -66,7 +103,9 @@ public partial class SignUp
         if (firstRender)
         {
             if (await TodoTemplateAuthenticationStateProvider.IsUserAuthenticated())
+            { 
                 NavigationManager.NavigateTo("/");
+            }
         }
     }
 }
