--- src/Tooling/Bit.Tooling.Templates/TodoTemplate/Web/Pages/SignUpConfirmation.razor.scss
@@ -1,21 +1,0 @@
-﻿@import '../Styles/abstracts/_functions.scss';
-@import '../Styles/abstracts/_media-queries.scss';
-
-.page-container {
-    display: flex;
-    flex-flow: column nowrap;
-    justify-content: center;
-    align-items: center;
-    width: 100%;
-    height: 100%;
-    padding: rem(56px) rem(16px);
-    box-sizing: border-box;
-
-    @include lg {
-        padding: rem(48px) rem(16px);
-    }
-
-    @include sm {
-        padding: rem(40px) rem(16px);
-    }
-}
